Day 20 part 2

+++ src/main/scala/domain/Coordinate.scala
@@ -1,3 +1,3 @@
 package domain
 
-case class Coordinate(num: Int)
+case class Coordinate(num: Long)
